O problema era overflow de Int32. O parquet atual tem:
text

total_qnt_vendas = Int32
soma sem cast = -541801479
soma com Int64 = 3753165817

Ou seja, o número real passa de 3,7 bilhões e estourava o limite do inteiro de 32 bits.

Corrigi em:

data_cache.py (line 604): novos caches de movimentação vão salvar total_qnt_vendas e total_qnt_sem_comprovacao como Int64.

analytics/dashboard.py (line 146): o dashboard agora força Int64 antes de somar.

services/dashboard.py (line 87): mesmo ajuste no serviço legado.

+++ backend/data_cache.py
@@ -601,8 +601,8 @@ def _sync_movimentacao(engine, progress_callback):
         pl.col("is_grande_rede").cast(pl.Boolean),
         pl.col("is_matriz").cast(pl.Boolean),
         pl.col("qtd_estabelecimentos_rede").cast(pl.Int64),
-        pl.col("total_qnt_vendas").cast(pl.Int32),
-        pl.col("total_qnt_sem_comprovacao").cast(pl.Int32),
+        pl.col("total_qnt_vendas").cast(pl.Int64),
+        pl.col("total_qnt_sem_comprovacao").cast(pl.Int64),
         pl.col("total_vendas").cast(pl.Float64),
         pl.col("total_sem_comprovacao").cast(pl.Float64),
     ]).sort(["id_cnpj", "periodo"])  # ORDENAÇÃO é a chave para compressão Parquet
